+++ tool/test.dart
@@ -16,7 +16,7 @@ void main(List<String> args) {
 
   dev.registerExtension('foo', _fooHandler);
 
-  dev.log('log from test');
+  // dev.log('log from test');
 
   // TODO: Handle this.
   dev.Timeline.timeSync('frame', _mockFrame);