+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fixed a NPE related to determining whether files are executable
 - fixed an exception when restoring saved settings
 - pass the route parameter to flutter launch configurations
+- add the ability to launch mojo apps from the flutter launch configuration
 
 ## 0.4.15
 - create hyperlinks from exception traces in the console view